+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 #### Changes
 - Return tournament statistics for all legs, not just the best per player
+- Add some additional properties for tournament overview
 
 #### Fixed
 - Always return legs won for player Shootout statistics